+++ lib/src/in_app_browser/android/in_app_browser_options.dart
@@ -11,13 +11,21 @@ class AndroidInAppBrowserOptions implements BrowserOptions, AndroidOptions {
   ///Set the action bar's title.
   String? toolbarTopFixedTitle;
 
-  ///Set to `false` to not close the InAppBrowser when the user click on the back button and the WebView cannot go back to the history. The default value is `true`.
+  ///Set to `false` to not close the InAppBrowser when the user click on the Android back button and the WebView cannot go back to the history. The default value is `true`.
   bool closeOnCannotGoBack;
+
+  ///Set to `false` to block the InAppBrowser WebView going back when the user click on the Android back button. The default value is `true`.
+  bool allowGoBackWithBackButton;
+
+  ///Set to `true` to close the InAppBrowser when the user click on the Android back button. The default value is `false`.
+  bool shouldCloseOnBackButtonPressed;
 
   AndroidInAppBrowserOptions(
       {this.hideTitleBar = false,
       this.toolbarTopFixedTitle,
-      this.closeOnCannotGoBack = true});
+      this.closeOnCannotGoBack = true,
+      this.allowGoBackWithBackButton = true,
+      this.shouldCloseOnBackButtonPressed = false});
 
   @override
   Map<String, dynamic> toMap() {
@@ -25,6 +33,8 @@ class AndroidInAppBrowserOptions implements BrowserOptions, AndroidOptions {
       "hideTitleBar": hideTitleBar,
       "toolbarTopFixedTitle": toolbarTopFixedTitle,
       "closeOnCannotGoBack": closeOnCannotGoBack,
+      "allowGoBackWithBackButton": allowGoBackWithBackButton,
+      "shouldCloseOnBackButtonPressed": shouldCloseOnBackButtonPressed,
     };
   }
 
@@ -33,6 +43,8 @@ class AndroidInAppBrowserOptions implements BrowserOptions, AndroidOptions {
     options.hideTitleBar = map["hideTitleBar"];
     options.toolbarTopFixedTitle = map["toolbarTopFixedTitle"];
     options.closeOnCannotGoBack = map["closeOnCannotGoBack"];
+    options.allowGoBackWithBackButton = map["allowGoBackWithBackButton"];
+    options.shouldCloseOnBackButtonPressed = map["shouldCloseOnBackButtonPressed"];
     return options;
   }