+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 5.2.1
 
 - Added `isRunning` method to the `HeadlessInAppWebView` class
+- Added `allowGoBackWithBackButton` and `shouldCloseOnBackButtonPressed` Android-specific InAppBrowser options
 - Fixed iOS `WebMessageListener` javascript implementation not calling event listeners when `onmessage` is set
 - Fixed `onCreateContextMenu` event on Android where `hitTestResult` has always `null` values
 - Fixed "java.lang.NullPointerException: Attempt to invoke virtual method 'void android.widget.SearchView.setQuery(java.lang.CharSequence, boolean)' on a null object reference" [#742](https://github.com/pichillilorenzo/flutter_inappwebview/issues/742)